+++ src/main/java/com/atlauncher/data/AbstractAccount.java
@@ -30,7 +30,6 @@ import javax.swing.ImageIcon;
 
 import com.atlauncher.FileSystem;
 import com.atlauncher.gui.dialogs.ProgressDialog;
-import com.atlauncher.gui.tabs.FeaturedPacksTab;
 import com.atlauncher.gui.tabs.InstancesTab;
 import com.atlauncher.gui.tabs.ServersTab;
 import com.atlauncher.managers.AccountManager;
@@ -68,8 +67,7 @@ public abstract class AbstractAccount implements Serializable {
     public String uuid;
 
     /**
-     * The pack names this account has collapsed in the {@link FeaturedPacksTab}, if
-     * any.
+     * The pack names this account has collapsed, if any.
      */
     public List<String> collapsedPacks = new ArrayList<>();