+++ src/main/java/com/atlauncher/gui/panels/packbrowser/FTBPacksPanel.java
@@ -106,6 +106,11 @@ public class FTBPacksPanel extends PackBrowserPlatformPanel {
     }
 
     @Override
+    public boolean supportsSearch() {
+        return true;
+    }
+
+    @Override
     public boolean hasCategories() {
         return false;
     }