+++ src/main/java/com/atlauncher/gui/panels/packbrowser/ModrinthPacksPanel.java
@@ -107,6 +107,11 @@ public class ModrinthPacksPanel extends PackBrowserPlatformPanel {
     }
 
     @Override
+    public boolean supportsSearch() {
+        return true;
+    }
+
+    @Override
     public boolean hasCategories() {
         return true;
     }