+++ src/main/java/com/atlauncher/gui/panels/packbrowser/TechnicPacksPanel.java
@@ -93,6 +93,11 @@ public class TechnicPacksPanel extends PackBrowserPlatformPanel {
     }
 
     @Override
+    public boolean supportsSearch() {
+        return true;
+    }
+
+    @Override
     public boolean hasCategories() {
         return false;
     }