+++ src/main/java/com/atlauncher/managers/PackManager.java
@@ -240,7 +240,6 @@ public class PackManager {
                     }
                     App.settings.addedPacks.add(packCode);
                     App.settings.save();
-                    App.launcher.refreshFeaturedPacksPanel();
                     App.launcher.refreshPacksBrowserPanel();
                     return true;
                 }
@@ -254,7 +253,6 @@ public class PackManager {
             if (Hashing.md5(code).equals(Hashing.toHashCode(packCode))) {
                 App.settings.addedPacks.remove(packCode);
                 App.settings.save();
-                App.launcher.refreshFeaturedPacksPanel();
                 App.launcher.refreshPacksBrowserPanel();
             }
         }