+++ CHANGELOG.md
@@ -11,5 +11,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Remove api key header from network debug logger
 - Show Fabric mods when browsing Modrinth mods and using Quilt loader [#575]
 - Servers not launching when there is a space in the directory path [#577]
+- Move Featured Packs tab to a sub tab within the Packs tab [#572]
 
 ### Misc